requirements.txt
asgiref==3.10.0
asttokens==3.0.0
brotli==1.2.0
certifi==2025.10.5
cfgv==3.4.0
charset-normalizer==3.4.4
colorama==0.4.6
crispy-tailwind==1.0.3
decorator==5.2.1
distlib==0.4.0
dj-database-url==3.0.1
Django==5.2.8
django-allauth==65.13.0
django-browser-reload==1.13.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-htmx==1.26.0
django-stubs==5.1.2
django-stubs-ext==5.2.7
django-tailwind==4.4.0
django-widget-tweaks==1.5.0
executing==2.2.1
filelock==3.20.0
gunicorn==23.0.0
identify==2.6.15
idna==3.11
iniconfig==2.3.0
ipython==8.28.0
jedi==0.19.2
Jinja2==3.1.6
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mypy==1.13.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
parso==0.8.5
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.0.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pure_eval==0.2.3
Pygments==2.19.2
pytailwindcss==0.3.0
pytest==8.4.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.32.5
ruff==0.8.5
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
text-unidecode==1.3
traitlets==5.14.3
types-pytz==2024.2.0.20241003
types-PyYAML==6.0.12.20250915
types-requests==2.32.0.20241016
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.35.4
wcwidth==0.2.14
whitenoise==6.11.0
PyJWT
cryptography
